Roles

+++ src/main/java/com/rose/procurement/purchaseOrder/entities/PurchaseOrderDto.java
@@ -1,20 +1,17 @@
 package com.rose.procurement.purchaseOrder.entities;
 
-import com.rose.procurement.category.entity.Category;
-import com.rose.procurement.contract.entities.Contract;
 import com.rose.procurement.enums.ApprovalStatus;
 import com.rose.procurement.enums.PaymentType;
 import com.rose.procurement.items.entity.Item;
-import com.rose.procurement.supplier.entities.Supplier;
 import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
-import java.util.function.Function;
 
 /**
  * DTO for {@link PurchaseOrder}
@@ -34,10 +31,10 @@ public class PurchaseOrderDto {
     @NotNull
     private PaymentType paymentType;
     private ApprovalStatus approvalStatus;
-    private double totalAmount;
-//    @NotNull
+    private BigDecimal totalAmount;
+    //    @NotNull
     private Set<Item> items;
-//    @NotNull
+    //    @NotNull
     private String vendorId;
 }
